requirement.txt
tensorflow 
numpy
Flask
Pillow